+++ CHANGELOG.md
@@ -12,7 +12,9 @@
   * Added the `api::projects::labels::PromoteLabel` endpoint to promote a project
     label to a group label.
   * Added the `api::projects:merge_requests::MergeMergeRequest` endpoint to
-    merge open merge requests
+    merge open merge requests.
+  * Added the `api::projects:merge_requests::RebaseMergeRequest` endpoint to
+    rebase open merge requests when using the fast-forward merge model.
 
 # v0.1301.1